Add tqdm library and file copy script using MVC pattern (#1)
* feat(dependencies): add tqdm to pyproject.toml dependencies

Tqdm provides a fast, extensible progress bar for loops and iterators. This addition will improve the user experience by giving visual feedback on the progress of long-running operations.

* feat: add file copy script using MVC pattern

The file copy script implements the Model-View-Controller design pattern. The JugglerModel class holds the source and destination paths, as well as the total size of the files being copied. The JugglerView class sets up a progress bar to show the progress of the copy operation. The JugglerController class performs the actual file copy operation and updates the progress bar in the JugglerView class after each chunk of data is written. The file's modification time, access time, and file permission are also preserved during the copy operation.